get conditions/add_conditions working
