+++ locales/en.json
@@ -246,15 +246,26 @@
   "conditions.title": "Section 5 - Patient's medical conditions",
   "conditions.intro": "List the condition that has the greatest functional impact first",
   "conditions.patient": "Patient:",
-  "conditions.name_of_condition": "Name of medical condition:",
-  "conditions.name_of_condition_hint": "ICD-9-CM code: XXXX",
-  "conditions.symptoms_began": "When did your patient's symptoms begin?",
-  "conditions.symptoms_began_hint": "For example, February 2008",
-  "conditions.clinically_impair": "How does this condition clinically impair your patient?",
-  "conditions.clinically_impair_hint": "For example: 'reduced range of motion', 'severe mood disturbance' etc",
-  "conditions.condition_outlook": "Is your patient's condition likely to:",
-  "condition.condition_outlook_improve": "Improve",
-  "condition.condition_outlook_deteriorate": "Deteriorate",
-  "condition.condition_outlook_staysame": "Stay the same",
-  "condition.condition_outlook_unknown": "Unknown"
+  "add_condition.name_of_condition": "Name of medical condition:",
+  "add_condition.name_of_condition_hint": "ICD-9-CM code: XXXX",
+  "add_condition.symptoms_began": "When did your patient's symptoms begin?",
+  "add_condition.symptoms_began_hint": "For example, February 2008",
+  "add_condition.clinically_impair": "How does this condition clinically impair your patient?",
+  "add_condition.clinically_impair_hint": "For example: 'reduced range of motion', 'severe mood disturbance' etc",
+  "add_condition.condition_outlook": "Is your patient's condition likely to:",
+  "add_condition.condition_outlook_improve": "Improve",
+  "add_condition.condition_outlook_deteriorate": "Deteriorate",
+  "add_condition.condition_outlook_staysame": "Stay the same",
+  "add_condition.condition_outlook_unknown": "Unknown",
+  "Section 1 - Get started": "Section 1 - Get started",
+  "Test ID": "Test ID",
+  "Test ID is required": "Test ID is required",
+  "Add condition": "Add condition",
+  "add_condition.condition_last": "How long do you expect the condition to last?",
+  "add_condition.condition_last_1_year": "Less than 1 year",
+  "add_condition.condition_last_more": "More than 1 year",
+  "add_condition.symptoms_occur": "How often do symptoms occur?",
+  "add_condition.symptoms_occur_periodically": "Periodically",
+  "add_condition.symptoms_occur_continuously": "Continuously",
+  "add_condition.symptoms_occur_unknown": "Unknown"
 }
