+++ routes/add_condition/js/add_condition.js
@@ -1,0 +1,1 @@
+// your js code here
