+++ routes/conditions/conditions.controller.js
@@ -1,12 +1,19 @@
-const { routeUtils } = require('./../../utils')
+const { routeUtils, getSessionData } = require('./../../utils')
 const { Schema } = require('./schema.js')
 
 module.exports = (app, route) => {
   const name = route.name
 
-  route.draw(app)
+  route
+    .draw(app)
     .get((req, res) => {
-      res.render(name, routeUtils.getViewData(req, {}))
+      const data = getSessionData(req)
+
+      if (!data.conditions) {
+        res.redirect('/en/add_condition')
+      } else {
+        res.render(name, routeUtils.getViewData(req, {}))
+      }
     })
     .post(route.applySchema(Schema), route.doRedirect())
 }
