+++ routes/personal/personal.njk
@@ -2,30 +2,12 @@
 
 {% block content %}
 
-    <h1>{{ __('Section 1 - Information about you') }}</h1>
-<example></example>
+    <h1>{{ __('Section 1 - Get started') }}</h1>
+
     <form method="post">
         <input type="hidden" name="_csrf" value="{{ csrfToken }}">
 
-        {{ textInput('social', 'Social Insurance Number', { hint: '443 856 299', class: "w-3-4", required: true }) }}
-
-        {{ radioButtons('preferred_title', { '1':'Mr.','2':'Mrs.','3':'Miss','4':'Ms.'}, data.preferred_title, 'Title', errors) }}
-        
-        {{ textInput('first_name', 'First name', { required: true }) }}
-        {{ textInput('middle_name', 'Middle name') }}
-        {{ textInput('last_name', 'Last name(s)', { required: true }) }}
-        {{ textInput('birth_last_name', 'Last name at birth (if different from above)') }}
-
-        {{ textInput('birthdate', 'Date of birth', { hint: 'YYYY-MM-DD', required: true }) }}
-
-        {{ textArea('address', 'Mailing address', { hint: '(No., Street, Apt., PO Box, RR), City/Town, Province/Territory, Country (if not Canada), Postal code', required: true}) }}
-
-        {{ textInput('telephone', 'Telephone number', { hint: 'XXX-XXX-XXXX', required: true }) }}
-        {{ textInput('alternate_telephone', 'Alternate telephone number', { hint: 'XXX-XXX-XXXX' }) }}
-
-        {{ radioButtons('contact_time', { 'morning':'Morning','afternoon':'Afternoon', 'letters_only':'Please don\'t call, send letters only'}, data.contact_time, 'The best time for Service Canada to contact you', errors, { required: true }) }}
-
-        <file-upload files="{{ data.files }}"></file-upload>
+        {{ textInput('name', 'Name', { class: "w-3-4", required: true }) }}
 
         {{ formButtons() }}
     </form>
